chore(ffi): rename `ciphersuite.rs` -> `cipher_suite.rs`
